requirements.txt
slack_sdk>=3.27.0
python-dotenv>=1.0
